requirements.txt


# torch
# pip install torch==2.10.0 torchvision==0.25.0 torchaudio==2.10.0 --index-url https://download.pytorch.org/whl/cu130
torchaudio
torchvision
Pillow
trimesh
meshio
matplotlib
numpy
scipy
polyscope
sounddevice
libigl
tqdm